third_party/WebKit/LayoutTests/fast/viewport/viewport-46-expected.txt

CONSOLE ERROR: line 3: The value for key "maximum-scale" is out of bounds and the value has been clamped.
CONSOLE ERROR: line 3: The value for key "maximum-scale" is out of bounds and the value has been clamped.
CONSOLE ERROR: line 3: The value for key "maximum-scale" is out of bounds and the value has been clamped.
ALERT: viewport size 32x35.2 scale 10 with limits [10, 10] and userScalable true
